### 2025-12-31 #9660 @crvernon
@hmcm  - Thank you for your submission to JOSS.  We are working through a large queue of submissions right now but will get you a topic editor assigned as soon as possible.
