+++ paper/paper.md
@@ -43,7 +43,7 @@ PyTupli addresses this gap by providing a dedicated Python toolkit for creating,
 
 Since the performance of offline RL algorithms often depends critically on the quality of the underlying dataset [@schweighofer2022dataset; @suttle2025behavioral; @asadulaev2025expert], PyTupli offers extensive filtering capabilities. Whereas established offline RL datasets primarily support filtering by entire episodes [@minari; @liu2023datasets], PyTupli enables tuple-level filtering as well. This can be used, for example, to rebalance datasets with sparse rewards or selectively include transitions from specific regions of the state space. In addition, PyTupli provides a suite of metrics for assessing dataset coverage and reward characteristics, which can serve as predictors of offline RL performance [@schweighofer2022dataset; @asadulaev2025expert] and help guide algorithm selection.
 
-# Related Software
+# State of the Field
 Publicly available tuple datasets have been essential for advancing offline RL algorithms [@kumar2020conservative; @kostrikov2021offline]. These curated collections span various domains, such as robotics and games [@fu2020d4rl; @minari; @formanek2023off; @gulcehre2020rl], power system control [@qin2022neorl], and autonomous driving [@liu2023datasets; @lee2024ad4rl], and are typically designed to support the development of improved offline RL methods. As offline RL techniques mature, they are being applied to increasingly diverse and task-specific control problems. Yet, to the best of our knowledge, no existing toolbox supports the collaborative creation, management, and sharing of datasets for custom environments. Minari [@minari] is the closest related tool, offering a repository of standardized datasets along with functionality for filtering, environment reconstruction, and recording new interactions. Its focus, however, remains on distributing datasets for established benchmarks.
 
 PyTupli instead targets collaborative workflows for custom control tasks. It enables researchers to share datasets and benchmarks directly within project teams without dependence on a central public server. Although Minari permits users to request publication on its official platform, this approach is often unsuitable for ongoing or proprietary work or for datasets that evolve over time. In addition, Minari does not provide tools for assessing dataset quality or coverage, which PyTupli includes to support informed dataset curation and algorithm selection in offline RL research.
