### 2026-02-24 #9660 @crvernon
@lockwo are you able to take this one on as editor?
