### 2026-02-25 #9660 @lockwo
There is a description of the software and it has a OSI approved license, so we can assemble the reviewers now.

@hmcm Please suggest ~5 potential reviewers. You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission. Or people who aren't in the JOSS system would also be ok. If you know their GitHub usernames, then please say them here in this thread (without tagging them with an @). You can also investigate the "AI"-suggested papers and their authors located in the previous posts from our editorial bot.
