### 2026-02-26 #9660 @lockwo
Hi 👋 @gvenugo3, @lorenzoschena, @ymahlau, @sandeshkatakam would two (or more) of you be willing to review this submission for JOSS?
