### 2026-03-02 #9660 @lockwo
Hi 👋 @janikmu, @ab93, @rahuldevikar, @cst0 would two (or more) of you be willing to review this submission for JOSS?
