### 2026-03-05 #10144 @editorialbot
[REVIEW]: PyTupli: Enabling Collaboration in Offline Reinforcement Learning
**Submitting author:** @hmcm (<a href="http://orcid.org/0009-0006-4603-9420">Hannah Markgraf</a>)
**Repository:** https://github.com/TUMcps/pytupli/
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.2.9
