### 2026-03-18 #10144 @rahuldevikar
Hi @hmcm
Raising some concerns:

1. The paper word count is 2112, exceeding the JOSS limit of 750–1750 words. The paper needs to be trimmed by approximately 360 words. The editorialbot flagged this
2.  Two claims in the Research Impact Statement need correction:
- 2.1 PyPI downloads: The paper claims "approximately 10,000 downloads." Publicly available PyPI Stats data (pypistats.org) shows ~65 downloads/month. The authors should either correct this number or provide methodology for how 10,000 was calculated (e.g., if it includes mirrors/CI downloads, this should be stated and contextualized).
- 2.2 CommonPower as "external" software: The paper describes CommonPower as "external research software," but it is maintained by the same team (same contributors, same TUMcps GitHub organization). Suggest rephrasing to accurately reflect the relationship (e.g., "companion project" or "related project by the same group").

3. 6 references have invalid DOIs as flagged by editorialbot
4. Only 2 of 6 paper authors (Markgraf, Eichelbeck) appear as GitHub contributors. Could the authors clarify what contributions the other four authors (Cappey, Demirtürk, Schattschneider, Althoff) made? This could be code contributions on GitLab, documentation, research design, supervision, etc.
5. The editorialbot flagged that 78.2% of code was added in a single 48-hour window. The GitHub repo has 0 issues, 0 PRs, and 0 releases. The 2026 JOSS guidelines include a "Development history and open-source practice" category. I'd welcome the authors' explanation of:
 - Whether there is additional development history on the GitLab LRZ instance that provides context
 - The project's plans for engaging with external users/contributors going forward
