+++ paper/paper.bib
@@ -40,7 +40,7 @@
   volume={33},
   pages={1179--1191},
   year={2020},
-  doi={10.5555/3495724.3495824}
+  url={https://dl.acm.org/doi/abs/10.5555/3495724.3495824}
 }
 @article{kostrikov2021offline,
   title={Offline reinforcement learning with implicit q-learning},
@@ -54,7 +54,7 @@
   author={Fu, Justin and Kumar, Aviral and Nachum, Ofir and Tucker, George and Levine, Sergey},
   journal={arXiv preprint arXiv:2004.07219},
   year={2020},
-  doi={https://doi.org/10.48550/arXiv.2004.07219}
+  url={https://doi.org/10.48550/arXiv.2004.07219}
 }
 @article{formanek2023off,
 author = {Formanek, Claude and Jeewa, Asad and Shock, Jonathan and Pretorius, Arnu},
@@ -63,7 +63,7 @@ year = {2023},
 publisher = {International Foundation for Autonomous Agents and Multiagent Systems},
 booktitle = {Proceedings of the 2023 International Conference on Autonomous Agents and Multiagent Systems},
 pages = {2442–2444},
-doi={10.5555/3545946.3598961}
+url={https://dl.acm.org/doi/abs/10.5555/3545946.3598961}
 }
 @article{gulcehre2020rl,
   title={Rl unplugged: A suite of benchmarks for offline reinforcement learning},
@@ -72,7 +72,7 @@ doi={10.5555/3545946.3598961}
   volume={33},
   pages={7248--7259},
   year={2020},
-  doi={10.5555/3495724.3496332}
+  url={https://dl.acm.org/doi/abs/10.5555/3495724.3496332}
 }
 
 @article{qin2022neorl,
@@ -82,7 +82,7 @@ doi={10.5555/3545946.3598961}
   volume={35},
   pages={24753--24765},
   year={2022},
-  doi={10.5555/3600270.3602065}
+  url={https://dl.acm.org/doi/abs/10.5555/3600270.3602065}
 }
 
 @inproceedings{lee2024ad4rl,
@@ -134,5 +134,5 @@ doi={10.5555/3545946.3598961}
   number={315},
   pages={1--20},
   year={2022},
-  doi={10.5555/3586589.3586904}
+  url={https://dl.acm.org/doi/abs/10.5555/3586589.3586904}
 }
