### 2026-03-19 32a2d6b
trimmed paper to match word count requirements
