+++ paper/paper.md
@@ -35,18 +35,21 @@ tags:
 Offline reinforcement learning (RL) offers a powerful way to derive effective decision-making policies for control problems using pre-collected data. However, managing and sharing such datasets in collaborative research settings can be challenging, as proper versioning, filtering, and access control are required. PyTupli is a free, Python-based toolkit designed to support this workflow by providing an easy-to-use yet specialized framework that remains independent of external service providers. Unlike centralized platforms, PyTupli is designed for self-hosted deployment, giving research teams full control over their data infrastructure and access policies. Its client library allows users to serialize and store control problems, upload new data, and retrieve precisely the subsets they need through flexible and expressive filters. Built-in metrics help evaluate dataset coverage and utility, informing both dataset selection and algorithm design for offline RL. To ensure secure deployment, a container-based server component offers authentication, role-based access control, and automated certificate provisioning. Together, these capabilities enable researchers to create, manage, exchange, and analyze datasets for offline RL in a robust and accessible manner.
 
 # Statement of need
-Reinforcement learning (RL) provides powerful methods for decision-making under uncertainty, but training RL agents typically requires extensive interaction with the underlying system or a computationally expensive simulator. Offline RL has emerged as a paradigm that alleviates this requirement by training agents solely on previously collected data [@lange2012batch]. Such datasets contain tuples consisting of *state, action, next state,* and *reward*, obtained from recordings of real systems or generated through simulators.
 
-Effectively managing, sharing, and curating these datasets is essential for collaborative offline RL research, yet existing tools provide only partial support. Platforms like Zenodo or the free version of HuggingFace allow users to share finalized datasets with a general public but are not suitable for ongoing or private collaborations. Furthermore, they lack mechanisms for tracking internal dataset structure or performing efficient, fine-grained queries. The same applies to version control systems such as GitHub. Traditional databases are better suited for this purpose but require substantial expertise to design and maintain robust workflows.
+Reinforcement learning (RL) provides powerful methods for decision-making under uncertainty, but training RL agents typically requires extensive interaction with the underlying system or a computationally expensive simulator. Offline RL alleviates this requirement by training agents solely on previously collected data [@lange2012batch]. Such datasets contain tuples of *state, action, next state,* and *reward*, obtained from real systems or simulators.
 
-PyTupli addresses this gap by providing a dedicated Python toolkit for creating, storing, and sharing tuple datasets for custom environments. We provide a Docker container that each research group or collaboration can deploy independently to maintain complete control over their data. Each dataset is associated to a benchmark. Benchmarks are stored as JSON-serialized objects and can be linked to related artifacts, including time-series data, algorithm hyperparameters, or trained policies, allowing multiple benchmarks to reference shared resources. Because it is built for scalable collaboration, PyTupli includes integrated user and access management features.
+Effectively managing, sharing, and curating these datasets is essential for collaborative offline RL research, yet existing tools provide only partial support. Platforms like Zenodo or the free version of HuggingFace allow users to share finalized datasets but are not suitable for ongoing or private collaborations. Furthermore, they lack mechanisms for tracking dataset structure or performing efficient queries. Traditional databases are better suited but require expertise to design and maintain workflows.
 
[…]
