+++ SECURITY.md
@@ -1,0 +1,26 @@
+# Security Policy
+
+## Supported versions
+
+This project is under active development. Security updates are applied to the latest version on the main branch.
+
+## Reporting a vulnerability
+
+If you discover a security vulnerability, please **do not open a public issue**.
+
+Instead, report it responsibly by emailing the maintainers (hannah.markgraf@tum.de).
+
+Please include:
+- A description of the vulnerability
+- Steps to reproduce
+- Potential impact
+- Suggested fixes (if any)
+
+## Response process
+
+We aim to:
+- Acknowledge reports within a few days
+- Investigate and confirm the issue
+- Provide a fix or mitigation as soon as possible
+
+We appreciate responsible disclosure and will credit contributors where appropriate.
