### 2026-03-23 #10144 @lockwo
Hi @rahuldevikar thanks for your quick review. Nothing pending on your side, once the PyTupl resolves your points raised, you can then iterate with them as needed or check off the items if they have been satisfactorily completed.
