+++ tests/server/test_api_users.py
@@ -29,7 +29,6 @@ async def test_token_creation_success(async_client):
     assert len(token_data['refresh_token']['token']) > 0
 
 
-
 async def test_token_creation_wrong_password(async_client):
     """Test token creation fails with wrong password."""
     response = await async_client.post(
@@ -40,7 +39,6 @@ async def test_token_creation_wrong_password(async_client):
     assert 'Incorrect username or password' in response.json()['detail']
 
 
-
 async def test_token_creation_nonexistent_user(async_client):
     """Test token creation fails with non-existent user."""
     response = await async_client.post(
@@ -49,7 +47,6 @@ async def test_token_creation_nonexistent_user(async_client):
     )
     assert response.status_code == 401
     assert 'Incorrect username or password' in response.json()['detail']
-
 
 
 async def test_token_creation_empty_credentials(async_client):
@@ -97,9 +94,10 @@ async def test_user_creation_success(async_client, admin_headers):
     assert 'memberships' in user_data
 
     # Cleanup
-    cleanup = await async_client.delete(f'/access/users/delete?username={username}', headers=admin_headers)
+    cleanup = await async_client.delete(
+        f'/access/users/delete?username={username}', headers=admin_headers
+    )
     assert cleanup.status_code == 200
-
 
 
 async def test_user_creation_duplicate_username(async_client, admin_headers):
@@ -128,7 +126,10 @@ async def test_user_creation_duplicate_username(async_client, admin_headers):
         assert 'already exists' in response.json()['detail']
     finally:
         # Cleanup
-        await async_client.delete(f'/access/users/delete?username={username}', headers=admin_headers)
+        await async_client.delete(
+            f'/access/users/delete?username={username}', headers=admin_headers
+        )
+
 
 # === USER LISTING TESTS ===
 
@@ -149,7 +150,6 @@ async def test_user_listing_success(async_client, admin_headers):
         assert 'username' in user
         assert 'memberships' in user
         assert 'password' not in user  # Password should never be returned
-
 
 
 async def test_user_listing_no_auth(async_client):
@@ -174,7 +174,9 @@ async def test_user_deletion_success(async_client, admin_headers):
     assert response.status_code == 200
 
     # Delete user
-    response = await async_client.delete(f'/access/users/delete?username={username}', headers=admin_headers)
+    response = await async_client.delete(
+        f'/access/users/delete?username={username}', headers=admin_headers
+    )
     assert response.status_code == 200
 
     # Verify user no longer exists in list
@@ -184,12 +186,12 @@ async def test_user_deletion_success(async_client, admin_headers):
     assert username not in usernames
 
 
-
 async def test_user_deletion_nonexistent(async_client, admin_headers):
     """Test deletion of non-existent user returns success (idempotent)."""
-    response = await async_client.delete('/access/users/delete?username=nonexistent_user_12345', headers=admin_headers)
+    response = await async_client.delete(
+        '/access/users/delete?username=nonexistent_user_12345', headers=admin_headers
+    )
     assert response.status_code == 200
-
 
 
 async def test_user_deletion_no_auth(async_client):
@@ -243,8 +245,9 @@ async def test_change_password_success(async_client, admin_headers):
 
     finally:
         # Cleanup
-        await async_client.delete(f'/access/users/delete?username={username}', headers=admin_headers)
-
+        await async_client.delete(
+            f'/access/users/delete?username={username}', headers=admin_headers
+        )
 
 
 async def test_change_password_nonexistent_user(async_client, admin_headers):
@@ -258,7 +261,6 @@ async def test_change_password_nonexistent_user(async_client, admin_headers):
     assert 'does not exist' in response.json()['detail']
 
 
-
[…]
