+++ tests/client/test_storage.py
@@ -294,27 +294,6 @@ def test_list_episodes_with_filter(storage, sample_episode_stored):
 
 
 @pytest.mark.parametrize('storage', STORAGE_TYPES, indirect=True)
-def test_list_episodes_with_nested_filter(
-    storage, sample_benchmark_stored, sample_episode_stored, sample_episode_2_stored
-):
-    """Test listing episodes with a nested filter."""
-    # Create a filter
-    episode_filter = FilterEQ(key='id', value=sample_episode_stored)
-    # benchmark id filter
-    benchmark_filter = FilterEQ(key='benchmark_id', value=sample_benchmark_stored)
-    # nested filter combining both
-    nested_filter = episode_filter & benchmark_filter
-    # Execute list_episodes with filter
-    episode_list = storage.list_episodes(filter=nested_filter)
-    # Verify results
-    assert episode_list is not None
-    assert isinstance(episode_list, list)
-    assert len(episode_list) == 1, 'Expected exactly one episode matching filter'
-    assert isinstance(episode_list[0], EpisodeHeader)
-    assert episode_list[0].metadata.get('agent') == 'test_agent'
-
-
-@pytest.mark.parametrize('storage', STORAGE_TYPES, indirect=True)
 def test_publish_episode(storage, sample_episode_stored):
     """Test publishing an episode."""
     episode_id = sample_episode_stored
