### 2026-03-23 7f844fd
fixed keyring.alt installation
