### 2026-03-24 #10144 @hmcm
Dear @rahuldevikar,

thank you very much for your time and your suggestions! To address your concerns, we have
- shortened the paper significantly,
- modified the Research Impact Statement as requested, and
- fixed the broken DOIs.

We have also addressed the issues raised on GitHub.

Regarding contributions of the other authors, Ms. Cappey, Ms. Demitürk, and Ms. Schattschneider developed an initial reduced version of this software as part of a student project from October 2024 - March 2025. Mr. Eichelbeck and I then extended and refactored the code, resulting in the present repository. Prof. Althoff has provided supervision over the course of the project.

The genesis of PyTupli is also why the majority of the code was created in the initial release. Since the original code was integrated into a different project, we believe that it would not be of much help to the reviewers. We apologize for this inconvenience!

To engage external users and contributors, PyTupli already features tutorials and detailed instructions for Docker-based deployment. Following your suggestions, we have also set up contribution guidelines, a code of conduct, and issue templates. To further grow the user and contributor base, we plan to actively advertise PyTupli within the reinforcement learning community, for example through presentations or demos at suitable workshops associated with the Reinforcement Learning Conference and the European Workshop on Reinforcement Learning, as well as via academic talks, collaborations, and targeted outreach to groups working on offline RL.
