### 2026-03-27 #10144 @rahuldevikar
Hi @hmcm,

Thank you for the quick and thorough response — really appreciate the detailed explanation on the authorship and development history. The student project origin (Oct 2024 – Mar 2025) makes the initial code dump make much more sense, and it's helpful to know Prof. Althoff's supervisory role.

A couple of things before I can check off the remaining items:

Changes not yet visible on GitHub: The community standards page (https://github.com/TUMcps/pytupli/community) still shows CONTRIBUTING, CODE_OF_CONDUCT, security policy, and issue templates as "Not added yet," and the main branch last commit is still "Bump version to 1.2.6" from 4 months ago. It looks like the changes may be on your GitLab instance but haven't been mirrored to the public GitHub repo yet. Could you push the latest to GitHub so I can verify?

Once the GitHub mirror is up to date, I'll verify:

- [x] Paper word count is within the 750–1750 limit
- [x]  Research Impact Statement has been corrected
- [x]  DOIs are fixed (will ask editorialbot to re-check)
- [x]  CONTRIBUTING.md, CODE_OF_CONDUCT.md, security policy, and issue templates are present
- [x]  GitHub Actions CI workflow is running

Regarding Issue #3 (GitHub Releases) — understood, that's fine. It was a non-blocking suggestion.

Thanks again for the responsiveness!
