### 2026-03-30 d666db6
Bump version to 1.2.7
