### 2026-04-05 33b7b40
Community guidelines, GitHub Actions workflow, linting
