+++ tests/client/test_client.py
@@ -1,15 +1,27 @@
 import pytest
 import uuid
 import keyring
-import time
-from unittest.mock import patch
 
 from pytupli.storage import TupliAPIClient, TupliStorageError
-from pytupli.schema import UserOut, UserRole, Group, GroupMembership, GroupMembershipQuery, GroupWithMembers
-from conftest import API_BASE_URL, API_USERNAME, API_PASSWORD, ADMIN_USERNAME, ADMIN_PASSWORD, admin_cleanup, clean_keyring
+from pytupli.schema import (
+    UserOut,
+    UserRole,
+    Group,
+    GroupMembership,
+    GroupMembershipQuery,
+    GroupWithMembers,
+)
+from conftest import (
+    API_BASE_URL,
+    API_USERNAME,
+    API_PASSWORD,
+    ADMIN_USERNAME,
+    ADMIN_PASSWORD,
+    admin_cleanup,
+)
 
 
-@pytest.fixture(scope="function")
+@pytest.fixture(scope='function')
 def api_client(clean_keyring):
     """Create a TupliAPIClient instance for testing."""
     client = TupliAPIClient()
@@ -21,11 +33,11 @@ def api_client(clean_keyring):
 def test_signup_success(api_client):
     """Test successful user signup."""
     # Create a unique test username to avoid conflicts
-    test_username = f"test_signup_{uuid.uuid4().hex[:8]}"
+    test_username = f'test_signup_{uuid.uuid4().hex[:8]}'
 
     try:
         # Call the signup method
-        user = api_client.signup(test_username, "test_password")
+        user = api_client.signup(test_username, 'test_password')
 
         # Verify results
         assert user is not None
@@ -41,15 +53,15 @@ def test_signup_success(api_client):
 def test_signup_user_exists(api_client):
     """Test signup when user already exists."""
     # Create a test user first
-    test_username = f"test_exists_{uuid.uuid4().hex[:8]}"
+    test_username = f'test_exists_{uuid.uuid4().hex[:8]}'
 
     try:
         # First signup should succeed
-        api_client.signup(test_username, "test_password")
+        api_client.signup(test_username, 'test_password')
 
         # Second signup with the same username should fail
-        with pytest.raises(TupliStorageError, match="Signup failed"):
-            api_client.signup(test_username, "different_password")
+        with pytest.raises(TupliStorageError, match='Signup failed'):
+            api_client.signup(test_username, 'different_password')
     finally:
         # Clean up the created user
         admin_cleanup(None, 'user', test_username)
@@ -85,8 +97,8 @@ def test_login_with_url(api_client, api_user):
 def test_login_invalid_credentials(api_client):
     """Test login with invalid credentials."""
     # Verify the correct exception is raised with wrong password
-    with pytest.raises(TupliStorageError, match="Login failed"):
-        api_client.login(API_USERNAME, "wrong_password")
+    with pytest.raises(TupliStorageError, match='Login failed'):
+        api_client.login(API_USERNAME, 'wrong_password')
 
     # Verify no tokens were stored
     assert keyring.get_password('pytupli', 'access_token') is None
@@ -131,16 +143,16 @@ def test_list_roles(api_user):
     assert all(isinstance(r, UserRole) for r in roles)
     # Check for expected roles
     role_names = [r.role for r in roles]
-    assert "admin" in role_names
-    assert "guest" in role_names
+    assert 'admin' in role_names
+    assert 'guest' in role_names
 
 
 def test_change_password(api_client):
     """Test successfully changing a user's password."""
     # Create a unique test user for this test
-    test_username = f"test_pwd_change_{uuid.uuid4().hex[:8]}"
-    initial_password = "initial_password"
-    new_password = "new_password_123"
[…]
