+++ CONTRIBUTING.md
@@ -1,0 +1,43 @@
+# Contributing
+
+Thank you for considering contributing to this project!
+
+## How to contribute
+
+There are several ways to contribute:
+- Reporting bugs
+- Suggesting features
+- Improving documentation
+- Submitting code changes
+
+## Reporting issues
+
+Please use GitHub Issues and include:
+- A clear description of the problem
+- Steps to reproduce
+- Expected vs actual behavior
+- Relevant logs, screenshots, or examples
+
+Before opening a new issue, please check if it already exists.
+
+## Pull requests
+
+To submit a contribution:
+1. Fork the repository and create a new branch
+2. Make your changes with clear, concise commits
+3. Ensure code is well-documented and tested (if applicable)
+4. Open a pull request with a clear description of your changes
+
+## Code style and guidelines
+
+- Follow existing code style and structure
+- Write clear, readable, and well-documented code
+- Add tests where appropriate
+
+## Code of conduct
+
+Please note that this project follows a Code of Conduct. Be respectful and constructive in all interactions.
+
+## Questions or support
+
+If you have questions, feel free to open an issue or contact the maintainers.

[…]
