+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pytupli"
-version = "1.2.7"
+version = "1.2.8"
 authors = [
     "Michael Eichelbeck <michael.eichelbeck@tum.de>",
     "Hannah Markgraf <hannah.markgraf@tum.de>",
