### 2026-04-12 #10144 @ab93
Hi @hmcm, overall the submission looks good. Would just like to flag a few things that I came across.

### Blocking

- The installation instructions for the end user is good with pip and poetry. However, it mixes up a bit with Local Development Setup with `poetry install`, and its optional dependencies. Generally these 2 sections are separate for a package, as they target different audiences.
- Please add the suggested DOIs raised by the editorial bot above.

### Non Blocking
- It would be helpful to pin the documentation link in the repo's About/description section for quick access.
- The main [doc page](https://pytupli.readthedocs.io/en/latest/index.html) does not seem to include the Deployment details, which I think is pretty important. It would be good for the end user to see everything in one go.
- There are some cells which have error outputs in the tutorial, e.g. [this](https://pytupli.readthedocs.io/en/latest/tutorials/Introduction.html#Uploading-and-Downloading-Benchmarks); would be nice to clean these up, and make the other cells a bit less verbose to make it easier for a user to follow them.

Feel free to ask any clarifications if you may have!
