### 2026-04-15 f8609c4
separate user and developer instructions
