+++ docs/source/index.rst
@@ -5,5 +5,6 @@ PyTupli - Tuple Intelligence for Offline RL
    :maxdepth: 2
 
    Quick-start <quickstart>
+   Deployment <deployment>
    Tutorials <tutorials>
    API reference <client_api>
