### 2026-04-16 #10144 @hmcm
Hi @ab93,
Thank you very much for your helpful suggestions! To address them, we
- split the README into two separate sections for "Installation" and "Development Setup"
- pinned the link to the documentation in the "About" section
- included deployment instructions in the documentation, and
- cleaned up the tutorial outputs.

Regarding the missing DOIs: I checked the two references in the paper, and they have valid DOIs. Is it possible that the editorial bot misflagged this?
Best
Hannah
