### 2026-04-23 #10144 @lockwo
The DOI missing is because the bot found what it thinks is a valid DOI, but the reference in the .bib doesn't have a DOI.

For example,

```
@article{formanek2023off,
author = {Formanek, Claude and Jeewa, Asad and Shock, Jonathan and Pretorius, Arnu},
title = {Off-the-Grid MARL: Datasets and Baselines for Offline Multi-Agent Reinforcement Learning},
year = {2023},
publisher = {International Foundation for Autonomous Agents and Multiagent Systems},
booktitle = {Proceedings of the 2023 International Conference on Autonomous Agents and Multiagent Systems},
pages = {2442–2444},
url={https://dl.acm.org/doi/abs/10.5555/3545946.3598961}
}
```

doesn't have a DOI, but 10.65109/vmil8420 seems like a correct DOI for it
