### 2026-04-29 #10144 @lockwo
Hi @ab93 thanks for the iteration. Is there more iterations you are thinking are necessary for the References (beyond the DOI fix) and the Collaborative effort component?
