+++ paper/paper.bib
@@ -63,7 +63,8 @@ year = {2023},
 publisher = {International Foundation for Autonomous Agents and Multiagent Systems},
 booktitle = {Proceedings of the 2023 International Conference on Autonomous Agents and Multiagent Systems},
 pages = {2442–2444},
-url={https://dl.acm.org/doi/abs/10.5555/3545946.3598961}
+url={https://dl.acm.org/doi/abs/10.5555/3545946.3598961},
+doi={10.65109/vmil8420}
 }
 @article{gulcehre2020rl,
   title={Rl unplugged: A suite of benchmarks for offline reinforcement learning},
@@ -82,7 +83,8 @@ url={https://dl.acm.org/doi/abs/10.5555/3545946.3598961}
   volume={35},
   pages={24753--24765},
   year={2022},
-  url={https://dl.acm.org/doi/abs/10.5555/3600270.3602065}
+  url={https://dl.acm.org/doi/abs/10.5555/3600270.3602065},
+  doi={10.52202/068431-1795}
 }
 
 @inproceedings{lee2024ad4rl,
