### 2026-05-29 b9efdb7
improved explanations of quality metrics, table, and tuple datasets
