+++ paper/paper.bib
@@ -34,7 +34,7 @@
   doi={10.48550/arXiv.2306.09303}
 }
 @article{kumar2020conservative,
-  title={Conservative q-learning for offline reinforcement learning},
+  title={Conservative {Q}-learning for offline reinforcement learning},
   author={Kumar, Aviral and Zhou, Aurick and Tucker, George and Levine, Sergey},
   journal={Advances in neural information processing systems},
   volume={33},
@@ -43,14 +43,14 @@
   url={https://dl.acm.org/doi/abs/10.5555/3495724.3495824}
 }
 @article{kostrikov2021offline,
-  title={Offline reinforcement learning with implicit q-learning},
+  title={Offline reinforcement learning with implicit {Q}-learning},
   author={Kostrikov, Ilya and Nair, Ashvin and Levine, Sergey},
   journal={arXiv preprint arXiv:2110.06169},
   year={2021},
   doi={10.48550/arXiv.2110.06169}
 }
 @article{fu2020d4rl,
-  title={D4rl: Datasets for deep data-driven reinforcement learning},
+  title={D4{RL}: Datasets for deep data-driven reinforcement learning},
   author={Fu, Justin and Kumar, Aviral and Nachum, Ofir and Tucker, George and Levine, Sergey},
   journal={arXiv preprint arXiv:2004.07219},
   year={2020},
@@ -67,7 +67,7 @@ url={https://dl.acm.org/doi/abs/10.5555/3545946.3598961},
 doi={10.65109/vmil8420}
 }
 @article{gulcehre2020rl,
-  title={Rl unplugged: A suite of benchmarks for offline reinforcement learning},
+  title={{RL} unplugged: A suite of benchmarks for offline reinforcement learning},
   author={Gulcehre, Caglar and Wang, Ziyu and Novikov, Alexander and Paine, Thomas and G{\'o}mez, Sergio and Zolna, Konrad and Agarwal, Rishabh and Merel, Josh S and Mankowitz, Daniel J and Paduraru, Cosmin and others},
   journal={Advances in Neural Information Processing Systems},
   volume={33},
@@ -129,7 +129,7 @@ doi={10.65109/vmil8420}
   doi={10.1109/SSCI50451.2021.9660006}
 }
 @article{seno2022d3rlpy,
-  title={d3rlpy: An offline deep reinforcement learning library},
+  title={{d}3rlpy: An offline deep reinforcement learning library},
   author={Seno, Takuma and Imai, Michita},
   journal={Journal of Machine Learning Research},
   volume={23},
