+++ paper/paper.md
@@ -20,7 +20,7 @@ authors:
 affiliations:
   - name: Technical University of Munich, Germany
     index: 1
-date: 10.12.2025
+date: 10 December 2025
 bibliography: paper.bib
 tags:
   - Python
@@ -36,7 +36,7 @@ Offline reinforcement learning (RL) offers a powerful way to derive effective de
 
 # Statement of need
 
-Reinforcement learning (RL) provides powerful methods for decision-making under uncertainty, but training RL agents typically requires extensive interaction with the underlying system or a computationally expensive simulator. Offline RL alleviates this requirement by training agents solely on previously collected data [@lange2012batch]. Such datasets contain tuples of *state, action, next state,* and *reward*, obtained from real systems or simulators.
+Reinforcement learning (RL) provides powerful methods for decision-making under uncertainty, but training RL agents typically requires extensive interaction with the underlying system or a computationally expensive simulator. Offline RL alleviates this requirement by training agents solely on previously collected data [@lange2012batch]. Such datasets contain tuples of state, action, next state, and reward, obtained from real systems or simulators, which we refer to as tuple datasets.
 
 Effectively managing, sharing, and curating these datasets is essential for collaborative offline RL research, yet existing tools provide only partial support. Platforms like Zenodo or the free version of HuggingFace allow users to share finalized datasets but are not suitable for ongoing or private collaborations. Furthermore, they lack mechanisms for tracking dataset structure or performing efficient queries. Traditional databases are better suited but require expertise to design and maintain workflows.
 
@@ -58,13 +58,13 @@ We briefly describe the core functionalities of PyTupli which are illustrated in
 
 **Benchmark and Artifact Management**: PyTupli enables users to store any control task defined as a gymnasium environment. Environments may include parameterizable configurations that produce variations in task dynamics. A fully specified environment is stored as a benchmark, providing a unique reference for reproducible evaluation of controllers. Benchmarks can reference additional data, such as exogenous inputs, time-series data, or pre-trained models. These external units, referred to as artifacts, are stored independently and linked to multiple benchmarks to avoid duplication.
 
-**Data Management**: PyTupli supports ingesting, storing, and querying structured datasets (RL tuples), including their relation to existing benchmark problems and any relevant metadata. MongoDB serves as the backend, providing scalable storage and retrieval. \autoref{tab:times} shows how ingestion and retrieval times scale with dataset size.
+**Data Management**: PyTupli supports ingesting, storing, and querying structured datasets (RL tuples), including their relation to existing benchmark problems and any relevant metadata. MongoDB serves as the backend, providing scalable storage and retrieval. \autoref{tab:times} shows how ingestion and retrieval times scale with dataset size. Since benchmarking was performed against a locally deployed server, network latency is excluded from all reported times and will add to these figures in practice.
 
 **Multi-User Collaboration and Access Control:** PyTupli facilitates collaborative workflows through private, group, and public scopes. Based on their assigned role, users can store, retrieve, delete, and publish objects. A server-side backend with FastAPI provides a REST interface for secure, programmatic access, while token-based authentication ensures secure sharing across teams or organizations.
 
 **Integration with Existing Offline RL Infrastructure:** An interface to the gymnasium framework enables users to record interactions with gymnasium environments as RL tuples. Furthermore, retrieved tuple datasets are made available in a form that can easily be converted into the dataset formats used by existing offline RL libraries such as d3rlpy [@seno2022d3rlpy].
 
-**Assessment of Dataset Quality:** PyTupli provides metrics for coverage and expected returns to guide dataset selection and algorithm choice.
[…]
