### 2026-05-31 #10144 @lockwo
@hmcm - please let me know if you see any issues with the proof that this generates
